fix: separate workers
