+++ packages/cf-worker/livestore-sync/wrangler.toml
@@ -1,7 +1,7 @@
-name = "geist-worker"
-main = "./src/index.ts"
+name = "geist-livestore-sync-worker"
+main = "../src/sync-worker.ts"
 compatibility_date = "2025-05-08"
-# compatibility_flags = ["nodejs_compat"]
+compatibility_flags = [ "nodejs_compat" ]
 
 secrets_store_secrets = [
   { binding = "STORACHA_AGENT_KEY_STRING", store_id= "d2267c5ce2614745a2735d712daf194e", secret_name = "STORACHA_AGENT_KEY_STRING" },
