+++ packages/cf-worker/src/sync-worker.ts
@@ -1,0 +1,34 @@
+import { type ExecutionContext, WorkerEntrypoint } from "cloudflare:workers";
+import {
+	type Env,
+	makeDurableObject,
+	makeWorker,
+} from "@livestore/sync-cf/cf-worker";
+
+export class WebSocketServer extends makeDurableObject({
+	onPush: async (message) => {
+		console.log("onPush", message.batch);
+	},
+	onPull: async (message) => {
+		console.log("onPull", message);
+	},
+}) {}
+
+// Note AutoRouter not compatabile
+const syncWorker = makeWorker({
+	validatePayload: (payload: any) => {
+		if (payload?.authToken !== "insecure-token-change-me") {
+			throw new Error("Invalid auth token");
+		}
+	},
+	enableCORS: true,
+});
+
+export class WorkerLiveStore extends WorkerEntrypoint {
+	env!: Env;
+	ctx!: ExecutionContext;
+	// Currently, entrypoints without a named handler are not supported
+	async fetch(request: Request) {
+		return syncWorker.fetch(request, this.env, this.ctx);
+	}
+}
